Changes to split populate in filter into two
