+++ src/Microsoft.OpenApi.CSharpAnnotations.DocumentGeneration/FilterSet.cs
@@ -88,8 +88,9 @@ namespace Microsoft.OpenApi.CSharpAnnotations.DocumentGeneration
                         _defaultFilterSet.Add(new SummaryToSummaryFilter());
 
                         //Pre processing operation filters
-                        _defaultFilterSet.Add(new ConvertAlternativeParamTagsFilter());
                         _defaultFilterSet.Add(new PopulateInAttributeFilter());
+                        _defaultFilterSet.Add(new ConvertAlternativeParamTagsFilter());
+                        _defaultFilterSet.Add(new ValidateInAttributeFilter());
                         _defaultFilterSet.Add(new BranchOptionalPathParametersFilter());
 
                         //Post processing document filters
