+++ src/Microsoft.OpenApi.CSharpAnnotations.DocumentGeneration/PreprocessingOperationFilters/PopulateInAttributeFilter.cs
@@ -88,29 +88,6 @@ namespace Microsoft.OpenApi.CSharpAnnotations.DocumentGeneration.PreprocessingOp
                     }
                 }
             }
-
-            paramElementsWithoutIn = paramElements.Where(p => p.Attribute(KnownXmlStrings.In)?.Value == null)
-                .ToList();
-
-            var paramElementsWithoutAllowedValues = paramElements.Where(
-                p => !KnownXmlStrings.AllowedInValues.Contains(p.Attribute(KnownXmlStrings.In)?.Value)).ToList();
-
-            if (paramElementsWithoutIn.Any())
-            {
-                throw new MissingInAttributeException(
-                    paramElementsWithoutIn.Select(
-                        p => p.Attribute(KnownXmlStrings.Name)?.Value));
-            }
-
-            if (paramElementsWithoutAllowedValues.Any())
-            {
-                throw new NotSupportedInAttributeValueException(
-                    paramElementsWithoutAllowedValues.Select(
-                        p => p.Attribute(KnownXmlStrings.Name)?.Value),
-                    paramElementsWithoutAllowedValues.Select(
-                        p => p.Attribute(KnownXmlStrings.In)?.Value)
-                );
-            }
         }
     }
 }
